+++ gerenciamento_hexagonal/application/Services/gerenciamento/meta.py
@@ -1,14 +1,14 @@
-from gerenciamento_hexagonal.application.services.gerenciamento.proposta import GerenciamentoPropostaServices
+from gerenciamento_hexagonal.application.services.interfaces.verify_gerenciamento import VerifyGerenciamentoExists
 from gerenciamento_hexagonal.domain.exceptions.gerenciamentoExceptions import NotFoundError
 from gerenciamento_hexagonal.domain.models.gerenciamento import GerenciamentoMeta
-from gerenciamento_hexagonal.domain.models.gerenciamentoDTO_Response import GerenciamentoMetaDTO
+from gerenciamento_hexagonal.domain.models.gerenciamento_dto_response import GerenciamentoMetaDTO
 from gerenciamento_hexagonal.domain.repositories.gerenciamento import GerenciamentoMetaRepository
 
 
 class GerenciamentoMetaServices:
-    def __init__(self, repository: GerenciamentoMetaRepository, service_proposta: GerenciamentoPropostaServices):
+    def __init__(self, repository: GerenciamentoMetaRepository, verify: VerifyGerenciamentoExists):
         self.repository = repository
-        self.service_proposta = service_proposta
+        self.verify = verify
 
     async def get_gerenciamento_meta(self) -> list[GerenciamentoMeta]:
         gerenciamento_metas = await self.repository.get_gerenciamento_meta()
@@ -23,7 +23,7 @@ class GerenciamentoMetaServices:
         return gerenciamento_meta
 
     async def create_gerenciamento_meta(self, gerenciamento_proposta_id: int, gerenciamento_meta: GerenciamentoMetaDTO) -> GerenciamentoMeta:
-        await self.service_proposta.get_gerenciamento_proposta_by_id(gerenciamento_proposta_id)
+        await self.verify.gerencimento_proposta_exists(gerenciamento_proposta_id)
 
         gerenciamento_meta = GerenciamentoMeta(**gerenciamento_meta.model_dump())
         gerenciamento_meta = await self.repository.create_gerenciamento_meta(gerenciamento_proposta_id, gerenciamento_meta)
@@ -39,7 +39,9 @@ class GerenciamentoMetaServices:
         return gerenciamento_meta
 
     async def delete_gerenciamento_meta(self, gerenciamento_meta_id: int) -> bool:
-        await self.get_gerenciamento_meta_by_id(gerenciamento_meta_id)
+        gerenciamento_meta = await self.get_gerenciamento_meta_by_id(gerenciamento_meta_id)
+        if gerenciamento_meta.arquivos_ids:
+            raise ValueError(f'gerenciament_meta with ID: {gerenciamento_meta_id} cannot be deleted because it has associated files with IDs: {gerenciamento_meta.arquivos_ids}.')
 
         await self.repository.delete_gerenciamento_meta(gerenciamento_meta_id)
